[UPDATE](regular/src/main/java/com/racacia/regular/model/vo/UrbanRuralVo.java): 修改城市乡村VO的ID和PID类型为Integer
更新了UrbanRuralVo和UrbanRural类中的id和pid字段类型，从Long更改为Integer，以增强数据的一致性和适应性。

[UPDATE](repository/src/main/java/com/racacia/repository/model/po/UrbanRural.java): 修改城市乡村PO的ID和PID类型为Integer

更新UrbanRural类中的id和pid字段类型，从Long更改为Integer，以提高数据存储的效率。

[UPDATE](regular/src/main/java/com/racacia/regular/web/UrbanRuralController.java): 增加根据PID查询城市信息的方法

增加了selectUrbanRuralByAreaClassIds方法，允许根据城市ID数组查询城市信息，增强了API功能。

[UPDATE](regular/src/main/java/com/racacia/regular/service/UrbanRuralService.java): 新增根据ID数组查询城市信息的方法

新增selectUrbanRuralByAreaClassIds方法，支持通过城市ID数组查询城市信息列表，提升服务灵活性。

[UPDATE](repository/src/main/java/com/racacia/repository/model/bo/UrbanRuralBo.java): 修改城市乡村BO的子项字段名称

将UrbanRuralBo中的urbanRuralBos字段更改为children，提升代码可读性。

[UPDATE](regular/src/main/java/com/racacia/regular/service/impl/UrbanRuralServiceImpl.java): 实现根据ID数组查询城市信息的方法

实现了selectUrbanRuralByAreaClassIds方法，支持根据城市ID数组查询并构建树形结构，增强了数据展示能力。

+++ repository/src/main/java/com/racacia/repository/model/po/UrbanRural.java
@@ -18,12 +18,12 @@ public class UrbanRural implements Serializable {
      * ID
      */
     @TableId(type = IdType.AUTO)
-    private Long id;
+    private Integer id;
 
     /**
      * 父ID
      */
-    private Long pid;
+    private Integer pid;
 
     /**
      * 代码
@@ -99,28 +99,28 @@ public class UrbanRural implements Serializable {
     /**
      * ID
      */
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
     /**
      * ID
      */
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
     /**
      * 父ID
      */
-    public Long getPid() {
+    public Integer getPid() {
         return pid;
     }
 
     /**
      * 父ID
      */
-    public void setPid(Long pid) {
+    public void setPid(Integer pid) {
         this.pid = pid;
     }
